backend/requirements.txt
google-analytics-data>=0.18.0
google-auth>=2.0.0
fastapi>=0.115.0
uvicorn>=0.34.0
openai>=1.30.0
python-dotenv>=1.0.0
gspread>=6.0.0
